test: add getClientPosition to assets
